overlord/swfeats: move ensure log test helper
