+++ overlord/ifacestate/ifacestate_test.go
@@ -60,6 +60,7 @@ import (
 	"github.com/snapcore/snapd/overlord/snapstate/snapstatetest"
 	"github.com/snapcore/snapd/overlord/state"
 	"github.com/snapcore/snapd/overlord/swfeats"
+	"github.com/snapcore/snapd/overlord/swfeats/swfeatstest"
 	"github.com/snapcore/snapd/release"
 	seccomp_compiler "github.com/snapcore/snapd/sandbox/seccomp"
 	"github.com/snapcore/snapd/snap"
@@ -12820,7 +12821,7 @@ func (s *interfaceManagerSuite) TestSystemKeyMismatchCompat(c *C) {
 }
 
 func (s *interfaceManagerSuite) TestEnsureLoopLogging(c *C) {
-	ensureLogs := testutil.CheckEnsureLoopLogging("ifacemgr.go", c, false)
+	ensureLogs := swfeatstest.CheckEnsureLoopLogging("ifacemgr.go", c, false)
 	knownEnsures := make(map[swfeats.EnsureEntry]bool)
 	for _, entry := range swfeats.KnownEnsures() {
 		knownEnsures[entry] = true
